Fixed compiler warning.
git-svn-id: svn+ssh://software.sandia.gov/svn/private/peridigm/trunk@1439 d9297710-12f9-482b-b11f-fb677dc88bbb

+++ src/core/Peridigm_NeighborhoodData.hpp
@@ -113,7 +113,7 @@ public:
 	return neighborhoodList;
   }
 
-  int const memorySize() const{
+  double const memorySize() const{
     int sizeInBytes =
       (2*numOwnedPoints + neighborhoodListSize + 2)*sizeof(int) + 3*sizeof(int*);
     double sizeInMegabytes = sizeInBytes/1048576.0;
